This commit solves two bugs in the non JS facet search:
DDBNEXT-974: Error in non javascript facet search
Before the facetValues were not url encoded. So special characters like
'(' were handled wrong.

DDBNEXT-984: Error in keyword selection for non js facet search
Using the mainUrl as a template for the subfacets urls makes problem. So
i replace it with an approach that is based on the request parameter
list